+++ src/main/java/com/groupe/cardgame/app/infrastructure/springboot/adapter/in/PullCardsController.java
@@ -1,16 +1,16 @@
-package com.groupe.cardgame.app.infra.springboot.adapter.in;
+package com.groupe.cardgame.app.infrastructure.springboot.adapter.in;
 
 import com.groupe.cardgame.app.application.PullCardsQuery;
 import com.groupe.cardgame.app.application.response.ApiResponse;
 import com.groupe.cardgame.app.application.response.ApiResponseWithBody;
-import com.groupe.cardgame.app.infra.springboot.adapter.in.exception.CardPackEmptyException;
-import com.groupe.cardgame.app.infra.springboot.adapter.in.exception.ResourceNotFoundException;
-import com.groupe.cardgame.app.infra.springboot.adapter.out.AddCardsToPlayerService;
-import com.groupe.cardgame.app.infra.springboot.adapter.out.CardPackRepository;
-import com.groupe.cardgame.app.infra.springboot.adapter.out.CardPullService;
-import com.groupe.cardgame.app.infra.springboot.adapter.out.PlayerRepository;
-import com.groupe.cardgame.app.infra.springboot.models.CardEntity;
-import com.groupe.cardgame.app.infra.springboot.models.CardPackEntity;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.in.exception.CardPackEmptyException;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.in.exception.ResourceNotFoundException;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.out.AddCardsToPlayerService;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.out.CardPackRepository;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.out.CardPullService;
+import com.groupe.cardgame.app.infrastructure.springboot.adapter.out.PlayerRepository;
+import com.groupe.cardgame.app.infrastructure.springboot.models.CardEntity;
+import com.groupe.cardgame.app.infrastructure.springboot.models.CardPackEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.nio.file.attribute.UserPrincipalNotFoundException;
 import java.util.ArrayList;
 import java.util.Locale;
 import java.util.Optional;
