+++ src/main/java/com/groupe/cardgame/app/infrastructure/springboot/adapter/in/exception/CardPackEmptyException.java
@@ -1,4 +1,4 @@
-package com.groupe.cardgame.app.infra.springboot.adapter.in.exception;
+package com.groupe.cardgame.app.infrastructure.springboot.adapter.in.exception;
 
 public class CardPackEmptyException  extends RuntimeException {
     public CardPackEmptyException(String message) {
