+++ src/main/java/com/groupe/cardgame/app/infrastructure/springboot/models/CardPackEntity.java
@@ -1,4 +1,4 @@
-package com.groupe.cardgame.app.infra.springboot.models;
+package com.groupe.cardgame.app.infrastructure.springboot.models;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -6,8 +6,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
