+++ src/main/java/com/groupe/cardgame/app/infrastructure/springboot/models/SpecialtyEntity.java
@@ -1,4 +1,4 @@
-package com.groupe.cardgame.app.infra.springboot.models;
+package com.groupe.cardgame.app.infrastructure.springboot.models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
